Unify RTC/millis time model, remove dead clock-drift persistence
Root-caused MonRing's erratic future-jumping event timestamps to
SimpleMeshTables.h passing RTC epoch-seconds where MonRing's appendEvent()
needed an already-resolved epoch-ms instant. Fixed by adding
RTCClock::nowMillis() as the single canonical millis-resolution clock
read (MeshCore.h), migrating every MonRing append/noteRadio/sampleEnv
call site onto it, and removing MonRing's own internal anchor-plus-
elapsed-millis() reconstruction (nowEpochMs()/_epochSecFromMillis()) now
that it's redundant. Packet::_rx_time and TuneController::tick() widened
to uint64_t/carry absolute epoch-ms accordingly.

Removed the persisted ahead-drift-offset mechanism
(beebo_clockDrift()/drift_off/applyDriftOffset_()) entirely: its write
side had gone unreachable in practice once Beebo::applyClockSync()'s
reboot-with-corrected-time redesign superseded it, and its own deferred-
log consumer was never actually called anywhere. Renamed the remaining
NVS timestamp-backup helper to setRebootRTCTime()/getRebootRTCTime() and
used the latter to replace ESP32RTCClock::begin()'s direct Preferences
read. Gave Beebo::handleCommand() its own "clock sync" branch (routed
through applyClockSync(), matching "clock"/"clock.epoch"/"time ") instead
of falling through to CommonCLI's seconds-only, repeater-role-only
version. Also removed the unused BEEBO_WIFI_OTA_STA web-OTA-over-STA
option (CommonCLI's own softAP path is what's actually reachable).

Native suite: 195/195. Hardware-verified on bench/bunch.

+++ src/Packet.h
@@ -66,7 +66,7 @@ public:
   // across a queued flood packet's delay wait even though other packets are
   // captured onto the ring in the meantime.
   uint32_t _rx_hash;    // SHA256(payload)[0:4] LE, distilled at capture
-  uint32_t _rx_time;    // capture-time clock reading, for the committed record's offset
+  uint64_t _rx_time;    // capture-time epoch-ms clock reading (RTCClock::nowMillis()), for the committed record's offset -- captured at RX, appendRx() may run meaningfully later (queued flood delay)
   int8_t   _rx_rssi;    // dBm at capture, for the ring record (independent of MSG_INCLUDE_RSSI)
   uint8_t  _rx_flags;   // nbr_len bits (see MonRing.h RXREC_FLAG_*)
   uint8_t  _rx_nbr[3];  // last-hop neighbour hash
